Suica版自動販売機問題リファクタ中
