+++ ruby/vending_machine/juice.rb
@@ -1,0 +1,8 @@
+class Juice
+  def initialize(name, price)
+    @name = name
+    @price = price
+  end
+
+  attr_reader :name, :price
+end
